fix: repair the 2026-09-08 bulk overwrite 14e1ce6d8 on the merits of each file
"Commit remaining workspace implementation and validation updates"
(14e1ce6d8, 555 files, +29,992/-54,368) was written from a checkout that
its own ledger (docs/issue-resolution-20260907.md) says predated 37
commits on main, and its notes say nothing was compiled or run. Measured
against its parent 0ef82308d by closest blob:

- 221 of its 385 modified files are byte-identical to versions of main
  from 73 to 426 commits earlier -> restored from 0ef82308d.
- 89 of its 93 deletions (bench/gha_results excluded) were files later
  commits had added: the source-removal guard, the public-API census, the
  assertionless-test detector and ledger, the test-census floor, every
  tests/bug_hunt_* pin, the examples smoke test, the thread-count pins,
  18 crate test modules -> restored.
- 33 renames moved the curated experiments/ investigations back to
  examples/, reversing 0eb45c231 -> reversed back.
- 4 "new" files were resurrections of files main had deleted with reasoned
  commits (the adaptive-regularization FD tests, the nuisance-atlas
  residue, the two #1412 throughput-gate tests) -> removed.
- it reverted the v0.3.155 version bump in 24 manifests, Cargo.lock,
  uv.lock and pyproject.toml and replaced the v0.3.155 and v0.3.154
  changelog entries with a 15-line "Unreleased" block -> restored.

The bulk commit's real work is kept: its 44 new files where their design
survives, the 92 modified files whose base was already main (the #2812
derived rho domains replacing the RhoBox hand boxes, the #979 fifth-order
tensors and third directional derivatives, the #1082 multinomial floor
removal, the tangent-precision refactor, the rayon psi-Gram refactor,
...), and every hunk that merged cleanly onto main.

Where the bulk commit and main both changed a file, every conflict was
read on both sides and decided on the code, not on which side landed:

- same fix on both sides (#979 missing-derivative refusal, #2731 exact-A
  carrier, #2598 predicate, the decrement-tolerance arm): main's superset.
- bulk edits to code main had deleted end to end (the O(n^-1) bias
  correction): no target, dropped.
- COMBINED: psi_hyper.rs (main's prepared-spectrum Jeffreys loop plus the
  bulk's explicit-curvature correction block); survival fit.rs (bulk's
  #2812 resolvability domain without the left-truncation floor main
  removed in a3b62b711, main's single authoritative outer run); marginal
  slope fit_setup.rs (bulk's caller-derived per-block domain, main's
  seed-centred domain as the fallback); payload builders (bulk's
  link-parameterised test helper, main's assertion folded in);
  atlas_topology.rs (bulk's nerve from row co-membership, keeping
  incoherent pairs in the unsigned nerve, plus main's intrinsic pile-up
  guard; both test modules, including the bulk's acceptance its author
  documents as failing until the atlas is repaired).
- the bulk's A.beta >= b documentation fix is correct (solver convention)
  and kept; its assert!/expect at the Python boundary reverse the
  panic-free boundary and are not.

Three parallel designs were set aside, each recorded with its reason:
#2828 exact beta-prior curvature (main's A_bb is FD-validated per term;
the bulk's ExactBetaRemainder carrier lacks main's #2267 and 711b96315
work, and its theta_bilinear third derivative is the one thing main may be
missing -- two probe tests are added to measure exactly that);
#2825 block dictionary (main's f32 floor was A/B-tested by the owner at
two bases with the suite green; the bulk's f64 transport reports its own
regression still failing); #2812/#2817 criterion resolution (the bulk's
note says it was never compiled and does not claim resolution; main's
decrement rung has measured receipts). The set-aside code is recoverable
from 14e1ce6d8 by path.

The full --all-targets check also found three breakages that exist on
main itself and are fixed here: a test reaching
firth_penalized_structural_rank through a module that does not exist
(3f9e1025d), and two tests still naming the deleted adaptive_regularization
and apply_bias_correction fields -- the second is why the quality_1082
binary could not build in Reference Quality Suite #572.

+++ crates/gam-sae/examples/wall_fair.rs
@@ -236,7 +236,7 @@ fn load_even_rows(
     path: &Path,
     take: usize,
     phase: u64,
-) -> Result<(Array2<f64>, Vec<usize>), String> {
+) -> Result<(Array2<f32>, Vec<usize>), String> {
     let mut file = File::open(path).map_err(|e| format!("open {}: {e}", path.display()))?;
     let header = read_npy_header(&mut file)?;
     let n_take = take.min(header.rows);
@@ -244,7 +244,7 @@ fn load_even_rows(
         .cols
         .checked_mul(std::mem::size_of::<f32>())
         .ok_or("row byte count overflow")?;
-    let mut out = Array2::<f64>::zeros((n_take, header.cols));
+    let mut out = Array2::<f32>::zeros((n_take, header.cols));
     let mut buf = vec![0u8; row_bytes];
     let stride = (header.rows / n_take.max(1)).max(1);
     let offset = (phase as usize) % stride;
@@ -257,7 +257,7 @@ fn load_even_rows(
             .map_err(|e| format!("read row {src_row}: {e}"))?;
         for c in 0..header.cols {
             let j = c * 4;
-            out[[i, c]] = f32::from_le_bytes([buf[j], buf[j + 1], buf[j + 2], buf[j + 3]]) as f64;
+            out[[i, c]] = f32::from_le_bytes([buf[j], buf[j + 1], buf[j + 2], buf[j + 3]]);
         }
     }
     Ok((out, vec![header.rows, header.cols]))
@@ -331,7 +331,7 @@ fn read_npy_header(file: &mut File) -> Result<NpyHeader, String> {
     })
 }
 
-fn peel_sink(x: &mut Array2<f64>, seed: u64, iterations: usize) -> f64 {
+fn peel_sink(x: &mut Array2<f32>, seed: u64, iterations: usize) -> f64 {
     let n = x.nrows();
     let p = x.ncols();
     let mut mean = vec![0.0f64; p];
@@ -346,7 +346,7 @@ fn peel_sink(x: &mut Array2<f64>, seed: u64, iterations: usize) -> f64 {
     let mut total = 0.0f64;
     for mut row in x.outer_iter_mut() {
         for c in 0..p {
-            row[c] -= mean[c] as f64;
+            row[c] -= mean[c] as f32;
             total += (row[c] as f64) * (row[c] as f64);
         }
     }
@@ -384,7 +384,7 @@ fn peel_sink(x: &mut Array2<f64>, seed: u64, iterations: usize) -> f64 {
     }
     for i in 0..n {
         for c in 0..p {
-            x[[i, c]] -= (scores[i] * v[c]) as f64;
+            x[[i, c]] -= (scores[i] * v[c]) as f32;
         }
     }
     sink_energy / total.max(1.0e-30)
@@ -397,7 +397,7 @@ fn normalize(v: &mut [f64]) {
     }
 }
 
-fn build_strata(x: &Array2<f64>) -> Vec<Stratum> {
+fn build_strata(x: &Array2<f32>) -> Vec<Stratum> {
     let mut bins: Vec<Vec<usize>> = vec![Vec::new(); 2048];
     let mut energies = vec![0.0f64; x.nrows()];
     for i in 0..x.nrows() {
@@ -474,7 +474,7 @@ fn sturges_cap(n: usize) -> usize {
 fn fit_stratum(
     label: &str,
     stratum: &Stratum,
-    x: &Array2<f64>,
+    x: &Array2<f32>,
     args: &Args,
 ) -> Result<Value, String> {
     let xs = take_rows(x, &stratum.rows);
@@ -589,8 +589,8 @@ fn block_config(n_blocks: usize, args: &Args) -> BlockSparseConfig {
     }
 }
 
-fn take_rows(x: &Array2<f64>, rows: &[usize]) -> Array2<f64> {
-    let mut out = Array2::<f64>::zeros((rows.len(), x.ncols()));
+fn take_rows(x: &Array2<f32>, rows: &[usize]) -> Array2<f32> {
+    let mut out = Array2::<f32>::zeros((rows.len(), x.ncols()));
     for (i, &row) in rows.iter().enumerate() {
         for c in 0..x.ncols() {
             out[[i, c]] = x[[row, c]];
@@ -599,7 +599,7 @@ fn take_rows(x: &Array2<f64>, rows: &[usize]) -> Array2<f64> {
     out
 }
 
-fn energy(x: &Array2<f64>) -> f64 {
+fn energy(x: &Array2<f32>) -> f64 {
     x.iter()
         .map(|&v| {
             let y = v as f64;
@@ -608,7 +608,7 @@ fn energy(x: &Array2<f64>) -> f64 {
         .sum()
 }
 
-fn energy_floor(target: &Array2<f64>, prediction: &Array2<f64>) -> f64 {
+fn energy_floor(target: &Array2<f32>, prediction: &Array2<f32>) -> f64 {
     let rss = target
         .iter()
         .zip(prediction.iter())
@@ -620,7 +620,7 @@ fn energy_floor(target: &Array2<f64>, prediction: &Array2<f64>) -> f64 {
     rss / energy(target).max(1.0e-30)
 }
 
-fn correction_proxy(base: &Array2<f64>, curved: &Array2<f64>, target_energy: f64) -> f64 {
+fn correction_proxy(base: &Array2<f32>, curved: &Array2<f32>, target_energy: f64) -> f64 {
     let e = base
         .iter()
         .zip(curved.iter())
